object s3 sink standardization
